refactor(runtime): split runtime.rs into focused submodules
Split the 1930-line astrid-runtime/src/runtime.rs monolith into a
runtime/ module directory with seven focused files:

- config.rs    (~65 lines)  — RuntimeConfig struct + constants
- mod.rs       (~310 lines) — struct, construction, session ops, accessors, helpers
- execution.rs (~255 lines) — run_turn_streaming, run_subagent_turn, run_loop
- tool_execution.rs (~320 lines) — execute_tool_call, execute_builtin_tool, execute_plugin_tool
- workspace.rs (~250 lines) — check_workspace_boundaries, path extraction helpers
- security.rs  (~200 lines) — FrontendApprovalHandler, classify_*, intercept_proof_to_auth_proof
- tests.rs     (~220 lines) — all 14 unit tests

No logic changes. Pure mechanical reorganisation with zero behavior change.
